+++ CHANGELOG.md
@@ -61,6 +61,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 * [#542](https://github.com/babylonlabs-io/finality-provider/pull/542) chore: abstract finality submitter
 * [#550](https://github.com/babylonlabs-io/finality-provider/pull/550) chore: use context for lifecycle
 * [#554](https://github.com/babylonlabs-io/finality-provider/pull/554) chore: rollup e2e parallel
+* [#555](https://github.com/babylonlabs-io/finality-provider/pull/555) chore: babylon controller use ctx
 
 ## v1.1.0-rc.1